removed selection provider as course provider does now this job (in accordance with the REST Api)
