+++ lib/connect/dataprovider/data.dart
@@ -1,6 +1,5 @@
 import 'package:courses_in_english/connect/dataprovider/campus/campus_provider.dart';
 import 'package:courses_in_english/connect/dataprovider/course/course_provider.dart';
-import 'package:courses_in_english/connect/dataprovider/course/selection_provider.dart';
 import 'package:courses_in_english/connect/dataprovider/department/department_provider.dart';
 import 'package:courses_in_english/connect/dataprovider/favorites/favorites_provider.dart';
 import 'package:courses_in_english/connect/dataprovider/lecturer/lecturer_provider.dart';
@@ -15,7 +14,6 @@ class Data {
   static final Data _instance = new Data._internal(new MockProviderFactory());
 
   CourseProvider courseProvider;
-  SelectionProvider selectionProvider;
   DepartmentProvider departmentProvider;
   LecturerProvider lecturerProvider;
   UserProvider userProvider;
@@ -31,7 +29,6 @@ class Data {
   /// Private default constructor
   Data._internal(ProviderFactory providerFactory)
       : courseProvider = providerFactory.createCourseProvider(),
-        selectionProvider = providerFactory.createSelectionProvider(),
         departmentProvider = providerFactory.createDepartmentProvider(),
         lecturerProvider = providerFactory.createLecturerProvider(),
         userProvider = providerFactory.createUserProvider(),
