+++ lib/connect/dataprovider/mock_provider_factory.dart
@@ -2,8 +2,6 @@ import 'package:courses_in_english/connect/dataprovider/campus/campus_provider.d
 import 'package:courses_in_english/connect/dataprovider/campus/mock/mock_campus_provider.dart';
 import 'package:courses_in_english/connect/dataprovider/course/course_provider.dart';
 import 'package:courses_in_english/connect/dataprovider/course/mock/mock_course_provider.dart';
-import 'package:courses_in_english/connect/dataprovider/course/mock/mock_selection_provider.dart';
-import 'package:courses_in_english/connect/dataprovider/course/selection_provider.dart';
 import 'package:courses_in_english/connect/dataprovider/department/department_provider.dart';
 import 'package:courses_in_english/connect/dataprovider/department/mock/mock_department_provider.dart';
 import 'package:courses_in_english/connect/dataprovider/favorites/favorites_provider.dart';
@@ -26,9 +24,6 @@ class MockProviderFactory implements ProviderFactory {
 
   @override
   LecturerProvider createLecturerProvider() => new MockLecturerProvider();
-
-  @override
-  SelectionProvider createSelectionProvider() => new MockSelectionProvider();
 
   @override
   UserProvider createUserProvider() => new MockUserProvider();
