+++ lib/connect/dataprovider/provider_factory.dart
@@ -1,6 +1,5 @@
 import 'package:courses_in_english/connect/dataprovider/campus/campus_provider.dart';
 import 'package:courses_in_english/connect/dataprovider/course/course_provider.dart';
-import 'package:courses_in_english/connect/dataprovider/course/selection_provider.dart';
 import 'package:courses_in_english/connect/dataprovider/department/department_provider.dart';
 import 'package:courses_in_english/connect/dataprovider/favorites/favorites_provider.dart';
 import 'package:courses_in_english/connect/dataprovider/lecturer/lecturer_provider.dart';
@@ -20,9 +19,6 @@ abstract class ProviderFactory {
 
   /// Create new instance of a department provider.
   DepartmentProvider createDepartmentProvider();
-
-  /// Create new instance of a selection provider.
-  SelectionProvider createSelectionProvider();
 
   /// Create new instance of a favorites provider.
   FavoritesProvider createFavoritesProvider();
