+++ lib/connect/dataprovider/sqlite_data.dart
@@ -1,6 +1,5 @@
 import 'package:courses_in_english/connect/dataprovider/campus/campus_provider.dart';
 import 'package:courses_in_english/connect/dataprovider/course/course_provider.dart';
-import 'package:courses_in_english/connect/dataprovider/course/selection_provider.dart';
 import 'package:courses_in_english/connect/dataprovider/department/department_provider.dart';
 import 'package:courses_in_english/connect/dataprovider/favorites/favorites_provider.dart';
 import 'package:courses_in_english/connect/dataprovider/lecturer/lecturer_provider.dart';
@@ -16,7 +15,6 @@ class SqliteData {
       new SqliteData._internal(new SqliteProviderFactory());
 
   CourseProvider courseProvider;
-  SelectionProvider selectionProvider;
   DepartmentProvider departmentProvider;
   LecturerProvider lecturerProvider;
   UserProvider userProvider;
@@ -32,7 +30,6 @@ class SqliteData {
   /// Private default constructor
   SqliteData._internal(ProviderFactory providerFactory)
       : courseProvider = providerFactory.createCourseProvider(),
-        selectionProvider = providerFactory.createSelectionProvider(),
         departmentProvider = providerFactory.createDepartmentProvider(),
         lecturerProvider = providerFactory.createLecturerProvider(),
         userProvider = providerFactory.createUserProvider(),
