+++ test/sql_test.dart
@@ -1,8 +1,7 @@
 //import 'package:courses_in_english/connect/dataprovider/campus/campus_provider.dart';
-import 'package:courses_in_english/connect/dataprovider/campus/mock/sqlite_campus_provider.dart';
+import 'package:courses_in_english/connect/dataprovider/campus/sqlite_campus_provider.dart';
 //import 'package:courses_in_english/connect/dataprovider/course/course_provider.dart';
-import 'package:courses_in_english/connect/dataprovider/course/mock/sqlite_course_provider.dart';
-import 'package:courses_in_english/connect/dataprovider/course/mock/mock_selection_provider.dart';
+import 'package:courses_in_english/connect/dataprovider/course/sqlite_course_provider.dart';
 import 'package:courses_in_english/connect/dataprovider/sqlite_data.dart';
 //import 'package:courses_in_english/connect/dataprovider/department/department_provider.dart';
 import 'package:courses_in_english/connect/dataprovider/department/sqlite_department_provider.dart';
@@ -45,7 +44,6 @@ void main() {
     expect(
         _data.departmentProvider, new isInstanceOf<SqliteDepartmentProvider>());
     expect(_data.favoritesProvider, new isInstanceOf<MockFavoritesProvider>());
-    expect(_data.selectionProvider, new isInstanceOf<MockSelectionProvider>());
     expect(_data.userProvider, new isInstanceOf<MockUserProvider>());
     expect(_data.campusProvider, new isInstanceOf<SqliteCampusProvider>());
   });
